pyproject.toml
[tool.poetry]
name = "hotels-telegram-bot"
version = "0.1.0"
description = "A bot which interacts with rapidapi's hotels api"
authors = ["Ostap Sukhovey <postrelovvv@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0"
requests = "^2.28.2"
pycountry = "^22.3.5"


[tool.poetry.group.dev.dependencies]
sphinx = "^6.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
